rename webRoutes to viewRoutes
